requirements.txt
django==3.2.4
gunicorn==20.1.0
psycopg2-binary==2.8.6
flake8==3.9.0
djangorestframework==3.11.0